pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kitty-voice-controller"
version = "0.1.0"
description = "Voice-controlled interface for Claude Code in Kitty terminal"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Stephen Sklarew"}
]
keywords = ["kitty", "terminal", "voice", "claude", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "openai-whisper>=20231117",
    "PyYAML>=6.0",
    "pynput>=1.7.6",
    "pyaudio>=0.2.14",
    "numpy>=1.24.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
claude-voice = "kitty_voice_controller.cli:main"

[project.urls]
Homepage = "https://github.com/stephensklarew/kitty-voice-controller"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
